feat: Memconfig client - demarcate clients into seperate packages && rename module to
