+++ e2e/server_test.go
@@ -5,7 +5,8 @@ import (
 	"testing"
 	"time"
 
-	"github.com/Layr-Labs/eigenda-proxy/client"
+	"github.com/Layr-Labs/eigenda-proxy/clients/memconfig_client"
+	"github.com/Layr-Labs/eigenda-proxy/clients/standard_client"
 	"github.com/Layr-Labs/eigenda-proxy/commitments"
 	"github.com/Layr-Labs/eigenda-proxy/common"
 	"github.com/Layr-Labs/eigenda-proxy/e2e"
@@ -68,10 +69,10 @@ func TestProxyClientServerIntegration(t *testing.T) {
 	ts, kill := e2e.CreateTestSuite(tsConfig)
 	t.Cleanup(kill)
 
-	cfg := &client.Config{
+	cfg := &standard_client.Config{
 		URL: ts.Address(),
 	}
-	daClient := client.New(cfg)
+	daClient := standard_client.New(cfg)
 
 	t.Run("single byte preimage set data case", func(t *testing.T) {
 		t.Parallel()
@@ -136,10 +137,10 @@ func TestProxyClient(t *testing.T) {
 	ts, kill := e2e.CreateTestSuite(tsConfig)
 	defer kill()
 
-	cfg := &client.Config{
+	cfg := &standard_client.Config{
 		URL: ts.Address(),
 	}
-	daClient := client.New(cfg)
+	daClient := standard_client.New(cfg)
 
 	testPreimage := e2e.RandBytes(100)
 
@@ -248,10 +249,10 @@ func TestProxyReadFallback(t *testing.T) {
 	ts, kill := e2e.CreateTestSuite(tsConfig)
 	defer kill()
 
-	cfg := &client.Config{
+	cfg := &standard_client.Config{
 		URL: ts.Address(),
 	}
-	daClient := client.New(cfg)
+	daClient := standard_client.New(cfg)
 	expectedBlob := e2e.RandBytes(1_000_000)
 	t.Log("Setting input data on proxy server...")
 	blobInfo, err := daClient.SetData(ts.Ctx, expectedBlob)
@@ -279,7 +280,7 @@ func TestProxyMemConfigClientCanGetAndPatch(t *testing.T) {
 	ts, kill := e2e.CreateTestSuite(tsConfig)
 	defer kill()
 
-	memClient := client.NewMemCfgClient(&client.Config{
+	memClient := memconfig_client.New(&memconfig_client.Config{
 		URL: "http://" + ts.Server.Endpoint(),
 	})
